Dashboard update
Added a dashboard sidebar to naavigate through current CRUD entities.
Fixed taskId and taskID naming errors.

Added searchbar functionality to each entity in dashboard.
Made the rows on each table clickable and opens a list of the children of clicked entity.
Made the URL in the dashboard change dynamically depending on what the admin clicks.

Fixed error handling in signup.