TaskModal connection
Connected the task modal to the board.jsx file so that task modals are interactable with tasks.
Fixed the wrong Ids in the controllers for creating board activities.
Made so changing labels/taskMembers, creating checklists/checklistItems and setting a duedate happen immediately. So the user does not have to wait for a token refresh